Convert configuration to use Viper
Having the only means of configuration be a config file that is baked
into the image means I need to rebuild the docker image to change
configuration. Viper is a fairly standard config tool that gives us an
easy way to add additional config mechanisms like environment variables
which make the image easier to configure via 12 factor methods.

This should make it easier to deploy as a container.