Add swagger json and UI to application
We're going to use this application as the base of a take home front-end
test in which the applicant is expected to develop a new front end for
the api. This means we need to much better document the API.

Add a swagger generator and all the tagging required to make that work,
and a swagger UI so that we can point users at the interface for the API
documentation.