Change docker entrypoint to simply the application
Having the ENTRYPOINT contain "serve" makes it more difficult than it
should be to do things like run the migrations or other commands that
are spawned from the application.

This is particularly obvious when you want to do something like
docker-compose where you want to specify the COMMAND like "updatedb" or
"serve" on top of the entrypoint.