Fix failing docker builds by building go-swagger
It appears that the downloaded go swagger is built with golang 1.10 and
installing it and running it against the golang 1.11 base causes a bunch
of problems.

We can download and build go-swagger ourselves, but unless we put it
into dep we can't pin a specific version and more and we're just running
from master.

That's a problem, but less of a problem than simply not building.